Refactor IBMTokenManager from utils
