+++ lithops/config.py
@@ -16,7 +16,6 @@
 #
 
 import os
-import sys
 import json
 import importlib
 import logging
