--- lithops/util/__init__.py
@@ -1,2 +1,0 @@
-from .ibm_token_manager import IBMTokenManager
-from .metrics import PrometheusExporter
